+++ src/Sharpliner/AzureDevOps/IfStringCondition.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using Sharpliner.AzureDevOps.ConditionedExpressions;
 using IfConditionOneOfArrayStringValue = Sharpliner.AzureDevOps.ConditionedExpressions.Arguments.IfConditionOneOfArrayStringValue;
 using IfConditionOneOfStringValue = Sharpliner.AzureDevOps.ConditionedExpressions.Arguments.IfConditionOneOfStringValue;
@@ -78,46 +76,4 @@ public abstract class IfStringCondition<T> : IfCondition<T>
     }
 
     internal override string Serialize() => WrapTag($"{_keyword}({WrapQuotes(_one)}, {WrapQuotes(_two)})");
-}
-
-internal class IfStringConditionHelper
-{
-    public static string Serialize(IfConditionOneOfStringValue value)
-    {
-        return value.Match(
-
-            str => str,
-            parameter => parameter.RuntimeExpression,
-            staticVariable => staticVariable.RuntimeExpression
-        );
-    }
-
-    public static string Serialize(IfConditionOneOfArrayStringValue arrayValue)
-    {
-        return arrayValue.Match(
-            strings => string.Join(", ", strings),
-            objects => Serialize(objects),
-            parameters => string.Join(", ", parameters.Select(p => Serialize(p))),
-            staticVariables => string.Join(", ", staticVariables.Select(v => Serialize(v)))
-        );
-    }
-
-    public static string Serialize(object[] array)
-    {
-        var convertedStringArray = array.Select(item =>
-            {
-                return item switch
-                {
-                    IfConditionOneOfStringValue oneOfStringValue => Serialize(oneOfStringValue),
-                    IfConditionOneOfArrayStringValue oneOfArrayStringValue => Serialize(oneOfArrayStringValue),
-                    StaticVariableReference staticVariableReference => Serialize(staticVariableReference),
-                    ParameterReference parameterReference => Serialize(parameterReference),
-                    VariableReference => throw new ArgumentException("If Conditions are compile-time statements, therefore runtime variables cannot be evaluated. You can use static variables or parameters instead."),
-                    _ => item.ToString()
-                };
-            })
-            .Select(Condition.WrapQuotes);
-
-        return string.Join(", ", convertedStringArray);
-    }
 }
