+++ src/Sharpliner/AzureDevOps/InlineStringCondition.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using Sharpliner.AzureDevOps.ConditionedExpressions;
 using InlineConditionOneOfArrayStringValue = Sharpliner.AzureDevOps.ConditionedExpressions.Arguments.InlineConditionOneOfArrayStringValue;
 using InlineConditionOneOfStringValue = Sharpliner.AzureDevOps.ConditionedExpressions.Arguments.InlineConditionOneOfStringValue;
@@ -77,45 +76,4 @@ public abstract class InlineStringCondition<T> : InlineCondition<T>
     }
 
     internal override string Serialize() => $"{_keyword}({WrapQuotes(_one)}, {WrapQuotes(_two)})";
-}
-
-internal class InlineStringConditionHelper
-{
-    public static string Serialize(InlineConditionOneOfStringValue value)
-    {
-        return value.Match(
-
-            str => str,
-            parameter => parameter.CompileTimeExpression,
-            variable => variable.RuntimeExpression
-        );
-    }
-
-    public static string Serialize(InlineConditionOneOfArrayStringValue arrayValue)
-    {
-        return arrayValue.Match(
-            strings => string.Join(", ", strings),
-            objects => Serialize(objects),
-            parameters => string.Join(", ", parameters.Select(p => Serialize(p))),
-            variables => string.Join(", ", variables.Select(v => Serialize(v)))
-        );
-    }
-
-    public static string Serialize(object[] array)
-    {
-        var convertedStringArray = array.Select(item =>
-            {
-                return item switch
-                {
-                    InlineConditionOneOfStringValue oneOfStringValue => Serialize(oneOfStringValue),
-                    InlineConditionOneOfArrayStringValue oneOfArrayStringValue => Serialize(oneOfArrayStringValue),
-                    ParameterReference parameterReference => Serialize(parameterReference),
-                    VariableReference variableReference => Serialize(variableReference),
-                    _ => item.ToString()
-                };
-            })
-            .Select(Condition.WrapQuotes);
-
-        return string.Join(", ", convertedStringArray);
-    }
 }
